requirements.txt
absl-py
numpy
tensorflow
tensorflow-datasets
tensorflow-probability